third_party/blink/web_tests/fast/html/imports/import-style-link-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
PASS childStyleLink.sheet.title is "Script Blocker"
PASS window.getComputedStyle(shouldBeBlue).color is "rgb(0, 0, 255)"
PASS window.getComputedStyle(shouldBeAqua).color is "rgb(0, 255, 255)"
PASS window.getComputedStyle(shouldBeYellow).color is "rgb(255, 255, 0)"
PASS window.getComputedStyle(shouldBeWhite).color is "rgb(255, 255, 255)"
PASS window.getComputedStyle(shouldBeTeal).color is "rgb(0, 128, 128)"
PASS window.getComputedStyle(shouldBeGreen).color is "rgb(0, 0, 0)"
PASS window.getComputedStyle(shouldBeGreen).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(shouldBeBlue).color is "rgb(0, 0, 0)"
PASS window.getComputedStyle(shouldBeGreen).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE

These elements should be styled appropriately:
Should be blue.
Should become aqua.
Should be white.
Should be yellow.
Should be teal.
Should become green.
